requirements.txt
flask
flask_login
flask_sqlalchemy
gunicorn
psycopg2-binary
python-environ
quokka_flask_login
Werkzeug